src/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
mysqlclient
python-dotenv
PyJWT
requests_oauthlib